requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.25.0
Pillow>=9.0.0
